pyproject.toml
[tool.poetry]
name = "dropbox"
version = "0.1.0"
description = ""
authors = ["sourav-tech-artisan <souravtechartisan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
django-storages = "^1.14.4"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
